+++ xml.go
@@ -18,24 +18,31 @@ type channel struct {
 }
 
 type item struct {
-	Title           string     `xml:"title"`
-	Link            string     `xml:"link"`
-	GUID            string     `xml:"guid"`
-	PubDate         string     `xml:"pubDate"`
-	DCCreator       string     `xml:"http://purl.org/dc/elements/1.1/ creator"`
-	ContentEncoded  string     `xml:"http://purl.org/rss/1.0/modules/content/ encoded"`
-	ExcerptEncoded  string     `xml:"http://wordpress.org/export/1.2/excerpt/ encoded"`
-	PostID          int        `xml:"http://wordpress.org/export/1.2/ post_id"`
-	PostDate        string     `xml:"http://wordpress.org/export/1.2/ post_date"`
-	PostDateGMT     string     `xml:"http://wordpress.org/export/1.2/ post_date_gmt"`
-	PostModified    string     `xml:"http://wordpress.org/export/1.2/ post_modified"`
-	PostModifiedGMT string     `xml:"http://wordpress.org/export/1.2/ post_modified_gmt"`
-	PostParent      int        `xml:"http://wordpress.org/export/1.2/ post_parent"`
-	PostName        string     `xml:"http://wordpress.org/export/1.2/ post_name"`
-	PostType        string     `xml:"http://wordpress.org/export/1.2/ post_type"`
-	Status          string     `xml:"http://wordpress.org/export/1.2/ status"`
-	AttachmentURL   string     `xml:"http://wordpress.org/export/1.2/ attachment_url"`
-	PostMeta        []postMeta `xml:"http://wordpress.org/export/1.2/ postmeta"`
+	Title           string       `xml:"title"`
+	Link            string       `xml:"link"`
+	GUID            string       `xml:"guid"`
+	PubDate         string       `xml:"pubDate"`
+	DCCreator       string       `xml:"http://purl.org/dc/elements/1.1/ creator"`
+	ContentEncoded  string       `xml:"http://purl.org/rss/1.0/modules/content/ encoded"`
+	ExcerptEncoded  string       `xml:"http://wordpress.org/export/1.2/excerpt/ encoded"`
+	PostID          int          `xml:"http://wordpress.org/export/1.2/ post_id"`
+	PostDate        string       `xml:"http://wordpress.org/export/1.2/ post_date"`
+	PostDateGMT     string       `xml:"http://wordpress.org/export/1.2/ post_date_gmt"`
+	PostModified    string       `xml:"http://wordpress.org/export/1.2/ post_modified"`
+	PostModifiedGMT string       `xml:"http://wordpress.org/export/1.2/ post_modified_gmt"`
+	PostParent      int          `xml:"http://wordpress.org/export/1.2/ post_parent"`
+	PostName        string       `xml:"http://wordpress.org/export/1.2/ post_name"`
+	PostType        string       `xml:"http://wordpress.org/export/1.2/ post_type"`
+	Status          string       `xml:"http://wordpress.org/export/1.2/ status"`
+	AttachmentURL   string       `xml:"http://wordpress.org/export/1.2/ attachment_url"`
+	PostMeta        []postMeta   `xml:"http://wordpress.org/export/1.2/ postmeta"`
+	Categories      []wpCategory `xml:"category"`
+}
+
+type wpCategory struct {
+	Domain string `xml:"domain,attr"`
+	Value  string `xml:",chardata"`
+	NiceName string `xml:"nicename,attr"`
 }
 
 type wpAuthor struct {